Parse.py: add an initial file that discovers commands/functions.
Add a new file called parse.py that reads a couple of files from a
trace32 installation and tries to figure out syntax/completions
for a possible REPL coming up in the future.

At the time of this writing, this file produces a high-quality set
of all trace32 functions, as well as their aliases and call
signatures.

Command-parsing is a much harder problem, and it's not 100% solved yet.
At the time of this writing, all commands are discovered, as well as
the known aliases of each command. Each command's various helpstrings
are also all located and added to the command's record. Helpstrings
aren't parsed for options/syntax yet, but that might come in later
commits.