Add new maintainer.mk file. Minor bugfixes.
Add a new maintainer.mk file that can be used to work with the
package without actually building/installing a wheel. The new file
also knows how to download, extract, and prepare the C API from
Lauterbach.com. This can be used for updating it to future versions
later down the road.